Use short timeout in creating socket, returning TimeoutError if creation timed out
If the connection is not possible, we want to fail fast. If we don't pass a timeout, the default behaviour is waiting up to 2s, which means this whole function takes > 6s even if the wait between retries is short.

In case we want to wait a while between retries to allow for a slow startup, this can be controlled by the retry_connection_after_s parameter.