Open logserver in new process group on windows
Otherwise, signals such as Ctrl+C are sent to both the main process and child process (the logger) together. However, we want to handle the termination of the logger ourselves, since we want to make sure it is up and running until the last moment. So, we open the logger in a separate group and deal with its termination ourselves in the shutdown of the main function.

With this addition, using ctrl+C to close the control room leads to a shutdown without any errors or irrelevant messages. Without it, the logserver will exit with a KeyboardInterrupt error, which is less clean, and adds unnecessary output.