Use waitress create_server instead of serve to enable proper shutdown
Waitress serve function is a convenience function that does not allow for easy lifetime management. Instead, by using create_server and run() we can call server.close() when we want to stop running the server and proceed to shutdown. To do so, we use `signal` to close the server when the programme receives a shutdown signal like SIGINT or SIGBREAK. This then closes the server, and the code proceeds to the finally block where the other components are gracefully shutdown.